Collection updater + tests refactoring
