+++ tests/Functional/Storage/Doctrine/PhpcrOdm/PhpcrOdmTestCase.php
@@ -4,6 +4,7 @@ namespace Psi\Component\ContentType\Tests\Functional\Storage\Doctrine\PhpcrOdm;
 
 use Doctrine\ODM\PHPCR\DocumentManagerInterface;
 use Psi\Component\ContentType\Tests\Functional\BaseTestCase;
+use Psi\Component\ContentType\Tests\Functional\Example\Storage\Doctrine\PhpcrOdm\Image;
 
 class PhpcrOdmTestCase extends BaseTestCase
 {
@@ -17,5 +18,16 @@ class PhpcrOdmTestCase extends BaseTestCase
         }
 
         $rootNode->addNode('test');
+    }
+
+    protected function createImage($path, $width, $height, $mimeType)
+    {
+        $image = new Image();
+        $image->path = $path;
+        $image->width = $width;
+        $image->height = $height;
+        $image->mimetype = $mimeType;
+
+        return $image;
     }
 }
